ml-service/requirements.txt
# ml-service/requirements.txt

torch
openai-whisper
fastapi
uvicorn[standard]
python-multipart